Ftr html address

+++ AeroBites/Controllers/AddressController.cs
@@ -19,9 +19,16 @@ namespace AeroBites.Controllers
             _addressService = addressService;
         }
 
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
-            return View();
+            var userId = User.GetId();
+            var addresses = await _context.Address
+                .Where(a => a.AccountId == userId)
+                .OrderByDescending(a => a.IsActive)
+                .ThenBy(a => a.Id)
+                .ToListAsync();
+
+            return View(addresses);
         }
 
         /// <summary>
@@ -33,13 +40,14 @@ namespace AeroBites.Controllers
         /// <param name="id">Identificador da conta.</param>
         /// <returns>Retorna o endereço adicionado.</returns>
         [HttpPost]
-        public async Task<IActionResult> AddAddress(double lat, double lng, string fullAddress, int id)
+        public async Task<IActionResult> AddAddress(double lat, double lng, string fullAddress)
         {
-            var address = await _addressService.AddAddressAccount(lat, lng, fullAddress, id);
+            var userId = User.GetId();
+            await _addressService.AddAddressAccount(lat, lng, fullAddress, userId);
 
             TempData[Enums.MessageType.successMessage.ToString()] = "Morada adicionada.";
 
-            return Ok(address);
+            return RedirectToAction("Index");
         }
 
         /// <summary>
@@ -50,13 +58,20 @@ namespace AeroBites.Controllers
         [HttpPost]
         public async Task<IActionResult> RemoveAddress(int id)
         {
+            var userId = User.GetId();  
             var address = await _context.Address.FindAsync(id);
+
+            if (address == null || address.AccountId != userId)
+            {
+                return NotFound();
+            }
+
             _context.Address.Remove(address);
             await _context.SaveChangesAsync();
 
             TempData[Enums.MessageType.successMessage.ToString()] = "Morada eliminada.";
 
-            return Ok();
+            return RedirectToAction("Index");
         }
 
         /// <summary>
@@ -84,6 +99,7 @@ namespace AeroBites.Controllers
         /// <returns>Retorna a lista de endereços do utilizador.</returns>
         public async Task<IActionResult> GetAllAddresses()
         {
+            var userId = User.GetId();
             var addressList = await _context.Address
                 .Where(address => address.AccountId == User.GetId())
                 .OrderByDescending(address => address.IsActive == true)
